Moved config utils
